feat: split leo3 feature surface into named gates
