+++ leo3/src/meta/expr.rs
@@ -53,7 +53,7 @@ impl LeanExpr {
     /// let var1 = LeanExpr::bvar(lean, 1)?;  // Next outer binding
     /// ```
     pub fn bvar<'l>(lean: Lean<'l>, idx: usize) -> LeanResult<LeanBound<'l, Self>> {
-        super::ensure_expr_initialized();
+        crate::runtime::ensure_expr_initialized();
         unsafe {
             let idx_nat = LeanNat::from_usize(lean, idx)?;
             let ptr = ffi::expr::lean_expr_mk_bvar(idx_nat.into_ptr());
@@ -118,7 +118,7 @@ impl LeanExpr {
         lean: Lean<'l>,
         level: LeanBound<'l, super::level::LeanLevel>,
     ) -> LeanResult<LeanBound<'l, Self>> {
-        super::ensure_expr_initialized();
+        crate::runtime::ensure_expr_initialized();
         unsafe {
             let ptr = ffi::expr::lean_expr_mk_sort(level.into_ptr());
             if ptr.is_null() {
@@ -144,7 +144,7 @@ impl LeanExpr {
         name: LeanBound<'l, LeanName>,
         levels: LeanBound<'l, LeanList>,
     ) -> LeanResult<LeanBound<'l, Self>> {
-        super::ensure_expr_initialized();
+        crate::runtime::ensure_expr_initialized();
         unsafe {
             let ptr = ffi::expr::lean_expr_mk_const(name.into_ptr(), levels.into_ptr());
             Ok(LeanBound::from_owned_ptr(lean, ptr))
@@ -166,7 +166,7 @@ impl LeanExpr {
         fn_expr: &LeanBound<'l, Self>,
         arg_expr: &LeanBound<'l, Self>,
     ) -> LeanResult<LeanBound<'l, Self>> {
-        super::ensure_expr_initialized();
+        crate::runtime::ensure_expr_initialized();
         unsafe {
             let lean = fn_expr.lean_token();
             // Increment reference counts since the FFI will store these pointers
@@ -249,7 +249,7 @@ impl LeanExpr {
         body: LeanBound<'l, Self>,
         binder_info: BinderInfo,
     ) -> LeanResult<LeanBound<'l, Self>> {
-        super::ensure_expr_initialized();
+        crate::runtime::ensure_expr_initialized();
         unsafe {
             let lean = binder_name.lean_token();
             let ptr = ffi::expr::lean_expr_mk_forall(
