+++ leo3/src/meta/mod.rs
@@ -116,92 +116,6 @@
 //! })?;
 //! ```
 
-use leo3_ffi as ffi;
-use std::sync::Once;
-
-// ============================================================================
-// Lazy Initialization
-// ============================================================================
-
-static PRELUDE_INIT: Once = Once::new();
-static EXPR_INIT: Once = Once::new();
-static ENV_INIT: Once = Once::new();
-
-/// Ensure Init.Prelude module is initialized (for Name functions)
-///
-/// This is called lazily when Name-related functions are first used.
-/// Safe to call multiple times - initialization happens only once.
-#[inline]
-pub(crate) fn ensure_prelude_initialized() {
-    PRELUDE_INIT.call_once(|| unsafe {
-        ffi::initialize_Init_Prelude(1, std::ptr::null_mut());
-    });
-}
-
-/// Ensure Lean.Expr module is initialized (for metaprogramming functions)
-///
-/// This is called lazily when Expr-related functions are first used.
-/// Safe to call multiple times - initialization happens only once.
-/// Note: This depends on Init.Prelude, so it initializes that first.
-///
-/// **Lean 4.20.0 Compatibility**: On macOS with Lean 4.20.0, there was a bug
-/// where `initialize_Lean_Expr` incorrectly initialized `l_Lean_levelZero` to 0x1
-/// (a tagged immediate) but `lean_level_mk_zero` expected it to be a heap pointer.
-/// This has been fixed by overriding `lean_level_mk_zero` in the FFI layer to
-/// return 0x1 directly, matching the behavior of Lean 4.27.0+.
-#[inline]
-pub(crate) fn ensure_expr_initialized() {
-    ensure_prelude_initialized(); // Expr depends on Prelude
-
-    EXPR_INIT.call_once(|| unsafe {
-        ffi::initialize_Lean_Expr(1, std::ptr::null_mut());
-    });
-}
-
-/// Ensure Lean.Environment module is initialized (for environment functions)
-///
-/// This is called lazily when Environment-related functions are first used.
-/// Safe to call multiple times - initialization happens only once.
-/// Note: This depends on Init.Prelude and Lean.Expr, so it initializes those first.
-/// `initialize_Lean_Environment` recursively initializes all transitive dependencies
-/// (each guarded by its own `_G_initialized` flag).
-#[inline]
-pub(crate) fn ensure_environment_initialized() {
-    ensure_expr_initialized(); // Environment depends on Expr
-
-    ENV_INIT.call_once(|| unsafe {
-        ffi::initialize_Lean_Environment(1, std::ptr::null_mut());
-        // Initialize Lean.Meta before marking end of initialization,
-        // since Meta functions (inferType, check) need their module
-        // globals set up during the initialization phase.
-        ffi::initialize_Lean_Meta(1, std::ptr::null_mut());
-        // Initialize the C++ utility module (name, expr infrastructure)
-        // and kernel module (type checker globals). Must be in this order,
-        // matching lean_initialize().
-        ffi::initialize_util_module();
-        ffi::initialize_kernel_module();
-        // Install default debug printer callback used by
-        // lean_expr_dbg_to_string.
-        ffi::init_default_print_fn();
-        ffi::initialize_library_core_module();
-        ffi::initialize_library_module();
-        // Mark initialization as complete so IO operations like
-        // lean_mk_empty_environment can proceed
-        ffi::lean_io_mark_end_initialization();
-    });
-}
-
-/// Ensure Lean.Meta module is initialized (for MetaM functions)
-///
-/// This is called lazily when MetaM-related functions are first used
-/// (e.g., `inferType`, `check`). Safe to call multiple times.
-/// Note: Meta initialization is handled inside `ensure_environment_initialized()`
-/// to ensure it runs before `lean_io_mark_end_initialization()`.
-#[inline]
-pub(crate) fn ensure_meta_initialized() {
-    ensure_environment_initialized();
-}
-
 pub mod context;
 pub mod declaration;
 pub mod environment;
