+++ leo3/tests/test_error_boundaries.rs
@@ -3,6 +3,8 @@
 //! Tests that Lean exceptions are properly caught, converted to Rust errors,
 //! and that the runtime remains usable after errors.
 
+#![cfg(feature = "meta")]
+
 use leo3::meta::*;
 use leo3::prelude::*;
 use leo3::KernelExceptionCode;
