+++ leo3/tests/test_literal_type_minimal.rs
@@ -1,4 +1,6 @@
 //! Test literal type
+#![cfg(feature = "meta")]
+
 use leo3::meta::*;
 use leo3::prelude::*;
