+++ leo3/tests/test_metam.rs
@@ -8,6 +8,8 @@
 //! (e.g., `inferType`) requires Phase 2 FFI bindings which aren't
 //! wired up yet. The `run()` tests here use synthetic computations.
 
+#![cfg(feature = "meta")]
+
 use leo3::meta::*;
 use leo3::prelude::*;
