+++ leo3/tests/test_prelude_integration.rs
@@ -7,6 +7,8 @@
 //! Since `LeanEnvironment::empty()` does not include the Lean prelude, we
 //! simulate prelude types by adding axioms for Nat, String, List, etc.
 
+#![cfg(feature = "meta")]
+
 use leo3::meta::*;
 use leo3::prelude::*;
