+++ leo3/tests/test_tactic.rs
@@ -10,6 +10,8 @@
 //! - Tactic failure (wrong type, unsolvable goal)
 //! - Tactic composition (chain multiple tactics)
 
+#![cfg(feature = "meta")]
+
 use leo3::meta::*;
 use leo3::prelude::*;
