+++ leo3/tests/test_type_checking.rs
@@ -4,6 +4,8 @@
 //! a variety of expression types. All tests use an empty environment (no prelude)
 //! so only Sort-based expressions are available.
 
+#![cfg(feature = "meta")]
+
 use leo3::meta::*;
 use leo3::prelude::*;
